more proof of concept (need to extract resources and format)

+++ app/src/main/java/com/smhvincent/moo_d/database/MoodDatabaseDao.kt
@@ -30,5 +30,5 @@ interface MoodDatabaseDao {
     fun getWithLiveData(key: Long): LiveData<Mood>
 
     @Query("SELECT AVG(rating) FROM mood_table")
-    suspend fun getRatingAverage(): Long
+    suspend fun getRatingAverage(): Double
 }
